# Conflicts:
#	.idea/modules.xml
#	gastos-services/src/main/java/com/example/gastos_service/controller/GastoController.java
#	gastos-services/src/main/java/com/example/gastos_service/service/GastoService.java
#	pom.xml

+++ categorias-services/src/main/java/com/example/categorias_services/service/CategoriaService.java
@@ -1,8 +1,12 @@
 package com.example.categorias_services.service;
 
+import com.example.categorias_services.client.GastoClient;
 import com.example.categorias_services.dto.CategoriaRequestDTO;
 import com.example.categorias_services.dto.CategoriaResponseDTO;
 import com.example.categorias_services.entity.Categoria;
+import com.example.categorias_services.exception.CategoriaMembershipException;
+import com.example.categorias_services.exception.CategoriaNotBelongToUser;
+import com.example.categorias_services.exception.CategoriaNotFoundException;
 import com.example.categorias_services.repository.CategoriaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,6 +20,8 @@ import java.util.stream.Collectors;
 public class CategoriaService {
 
     private final CategoriaRepository categoriaRepository;
+    private final GastoClient gastoClient;
+    private final TokenService tokenService;
 
     public CategoriaResponseDTO criarCategoria(String usuarioEmail, CategoriaRequestDTO dto) {
         Categoria categoria = Categoria.builder()
@@ -37,10 +43,10 @@ public class CategoriaService {
 
     public CategoriaResponseDTO buscarPorId(UUID id, String usuarioEmail) {
         Categoria categoria = categoriaRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Categoria não encontrada"));
+                .orElseThrow(() -> new CategoriaNotFoundException("Categoria não encontrada"));
 
         if (!categoria.getUsuarioId().equals(usuarioEmail)) {
-            throw new RuntimeException("Categoria não pertence ao usuário");
+            throw new CategoriaNotBelongToUser("Categoria não pertence ao usuário");
         }
 
         return toResponseDTO(categoria);
@@ -48,10 +54,15 @@ public class CategoriaService {
 
     public void deletar(UUID id, String usuarioEmail) {
         Categoria categoria = categoriaRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Categoria não encontrada"));
+                .orElseThrow(() -> new CategoriaNotFoundException("Categoria não encontrada"));
 
         if (!categoria.getUsuarioId().equals(usuarioEmail)) {
-            throw new RuntimeException("Categoria não pertence ao usuário");
+            throw new CategoriaNotBelongToUser("Categoria não pertence ao usuário");
+        }
+
+        String token = tokenService.getCurrentToken();
+        if (gastoClient.existeGastoComCategoria(id, token)) {
+            throw new CategoriaMembershipException("Não é possível excluir a categoria pois existem gastos associados a ela");
         }
 
         categoriaRepository.deleteById(id);
